bug 1490727: Move to contributions app
