+++ kuma/contributions/jinja2/contributions/contribute.html
@@ -30,6 +30,6 @@
 
 {% block document_head %}
     {% with form=form, is_cta=False %}
-        {% include "landing/includes/contribution-form.html" %}
+        {% include "contributions/includes/contribution-form.html" %}
     {% endwith %}
 {% endblock %}
